requirements.txt
pandas
azure-storage-blob
psycopg2-binary
apache-airflow